fix: rename levels
