+++ src/core/game-figures/easy/camel.js
