+++ src/core/game-figures/easy/celtic-cross.js
