+++ src/core/game-figures/easy/location-mark.js
