+++ src/core/game-figures/easy/top-hat.js
