+++ src/core/game-figures/easy/tower.js
