+++ src/core/game-figures/hard/anchor.js
