+++ src/core/game-figures/hard/death.js
