+++ src/core/game-figures/hard/fish.js
