+++ src/core/game-figures/hard/horse.js
