+++ src/core/game-figures/hard/ram.js
