+++ src/core/game-figures/medium/bird.js
