+++ src/core/game-figures/medium/cat.js
