+++ src/core/game-figures/medium/deer.js
