+++ src/core/game-figures/medium/duck.js
