+++ src/core/game-figures/medium/house.js
