+++ src/core/game-figures/medium/snowman.js
